rename to lammps_data, so we can have a lammps_traj parser
